Optimize loading speed of multiple levels. Profiling in debug build.
A flag is set in window when loading more than one level that specifies
whether we are in the middle of loading a level and prevents on_switch_page
from updating any tilesets when it triggers except for the last level to be
loaded.
Additionally several functions with update_all side-effect take an additional
parameter to overwrite that behavior.