Allow floating point NPC positions
npc::set_level_x/y round the passed in float to 0.5 accuracy and store the
result as an integer. npc::get_level_x/y return the position as a float by
converting the integer back.
This also fixes an infinite loop when loading a level with NPC fractional
NPC positions.
Also fixed the creation of an undo NPC, previously the drag start position
was not properly set.